+++ src/plugins/screenshot/windows/capture/capture-preload.ts
