+++ src/plugins/screenshot/windows/capture/capture.html
@@ -49,7 +49,7 @@
             tagName: "audio",
             id: "screenshotAudio",
         });
-        screenshotAudio.src = "../audio/camera_shutter.m4a";
+        screenshotAudio.src = "../../audio/camera_shutter.m4a";
 
         const h1 = createElement({
             tagName: "h1",
