+++ src/plugins/screenshot/windows/capture/style.css
