+++ src/plugins/screenshot/windows/crop/crop-preload.ts
