+++ src/plugins/screenshot/windows/crop/crop.css
