+++ src/plugins/screenshot/windows/crop/crop.html
