+++ src/plugins/screenshot/windows/preview/preview-preload.ts
@@ -1,6 +1,6 @@
 import { PostageBatch } from '@ethersphere/bee-js'
 import { contextBridge, ipcRenderer } from 'electron'
-import type { OnImageDataURL } from '../interface'
+import type { OnImageDataURL } from '../../interface'
 
 contextBridge.exposeInMainWorld('electron', {
   onImageDataURL: (cb: OnImageDataURL) => ipcRenderer.on('image-data-url', (_, dataURL) => cb(dataURL)),
