+++ src/plugins/screenshot/windows/preview/preview.css
